HeartRisk Predict/HeartRisk-Prediction-MLOps-Streamlit-Docker-main/HeartRisk/requirements.txt
ansicon==1.89.0
asgiref==3.8.1
blessed==1.20.0
catalogue==2.0.10
contexttimer==0.3.3
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
gitdb==4.0.11
GitPython==3.1.43
humanfriendly==10.0
jinxed==1.2.1
joblib==1.3.2
npm==0.1.1
numpy==1.26.4
nvm==1.0.21
optional-django==0.1.0
pandas==2.2.1
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.4.1.post1
scipy==1.13.0
six==1.16.0
smmap==5.0.1
sqlparse==0.4.4
srsly==2.4.8
threadpoolctl==3.4.0
tzdata==2024.1
tzlocal==5.2
wcwidth==0.2.13
xgboost==1.5.0
